+++ changelog.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 
 ## [Unreleased]
 
+- Added profile pictures: users now can select and crop an image and set it as their profile picture
 - Allow theming the UI's colors using the platform's design table 
 
 ## [Released]